SCREAM: split impl file of P3Microphysics to remove ICE with nvcc
Putting run_impl in a separate cpp file seems to remove the
internal compiler error (ICE).
