+++ components/scream/src/physics/p3/atmosphere_microphysics.cpp
@@ -11,9 +11,6 @@
 
 namespace scream
 {
-/*
- * P3 Microphysics routines
-*/
 
   using namespace p3;
 
@@ -273,41 +270,6 @@ void P3Microphysics::initialize_impl (const util::TimeStamp& /* t0 */)
   // -- Set values for the post-amble structure
   p3_postproc.set_variables(m_num_cols,nk_pack,prog_state.th,pmid,T_atm,t_prev,prog_state.qv,qv_prev,
       diag_outputs.diag_eff_radius_qc,diag_outputs.diag_eff_radius_qi);
-}
-
-// =========================================================================================
-void P3Microphysics::run_impl (const Real dt)
-{
-
-  // Assign values to local arrays used by P3, these are now stored in p3_loc.
-  Kokkos::parallel_for(
-    "p3_main_local_vals",
-    Kokkos::RangePolicy<>(0,m_num_cols),
-    p3_preproc
-  ); // Kokkos::parallel_for(p3_main_local_vals)
-  Kokkos::fence();
-
-  // Update the variables in the p3 input structures with local values.
-
-  infrastructure.dt = dt;
-  infrastructure.it++;
-
-  // WorkspaceManager for internal local variables
-  const Int nk_pack = ekat::npack<Spack>(m_num_levs);
-  const auto policy = ekat::ExeSpaceUtils<KT::ExeSpace>::get_default_team_policy(m_num_cols, nk_pack);
-  ekat::WorkspaceManager<Spack, KT::Device> workspace_mgr(m_buffer.wsm_data, nk_pack, 52, policy);
-
-  // Run p3 main
-  P3F::p3_main(prog_state, diag_inputs, diag_outputs, infrastructure,
-               history_only, workspace_mgr, m_num_cols, m_num_levs);
-
-  // Conduct the post-processing of the p3_main output.
-  Kokkos::parallel_for(
-    "p3_main_local_vals",
-    Kokkos::RangePolicy<>(0,m_num_cols),
-    p3_postproc
-  ); // Kokkos::parallel_for(p3_main_local_vals)
-  Kokkos::fence();
 }
 
 // =========================================================================================
